+++ src/components/LegacySidebarContext.tsx
@@ -5,12 +5,10 @@ interface SidebarState {
   toggled?: boolean;
   broken?: boolean;
   rtl?: boolean;
-  width?: string;
-  collapsedWidth?: string;
   transitionDuration?: number;
 }
 
-export interface SidebarContextProps extends SidebarState {
+export interface LegacySidebarContextProps extends SidebarState {
   updateSidebarState: (values: SidebarState) => void;
   updateCollapseState: () => void;
   updateToggleState: () => void;
@@ -20,10 +18,18 @@ interface SidebarProviderProps {
   children?: React.ReactNode;
 }
 
-export const SidebarContext = React.createContext<SidebarContextProps | undefined>(undefined);
+export const LegacySidebarContext = React.createContext<LegacySidebarContextProps | undefined>(
+  undefined,
+);
 
 export const SidebarProvider: React.FC<SidebarProviderProps> = ({ children }) => {
-  const [sidebarState, setSidebarState] = React.useState<SidebarState>();
+  const [sidebarState, setSidebarState] = React.useState<SidebarState>({
+    collapsed: false,
+    toggled: false,
+    broken: false,
+    rtl: false,
+    transitionDuration: 300,
+  });
 
   const updateSidebarState = React.useCallback((values: Partial<SidebarState>) => {
     setSidebarState((prevState) => ({ ...prevState, ...values }));
@@ -42,5 +48,7 @@ export const SidebarProvider: React.FC<SidebarProviderProps> = ({ children }) =>
     [sidebarState, updateCollapseState, updateSidebarState, updateToggleState],
   );
 
-  return <SidebarContext.Provider value={providerValue}>{children}</SidebarContext.Provider>;
+  return (
+    <LegacySidebarContext.Provider value={providerValue}>{children}</LegacySidebarContext.Provider>
+  );
 };