+++ src/components/Menu.tsx
@@ -38,8 +38,22 @@ export interface MenuContextProps {
    * @default ```300```
    */
   transitionDuration?: number;
+
+  /**
+   * If set to true, the popper menu will close when a menu item is clicked
+   * This works on collapsed mode only
+   * @default ```false```
+   */
   closeOnClick?: boolean;
+
+  /**
+   * Apply styles to MenuItem and SubMenu components and their children
+   */
   menuItemStyles?: MenuItemStyles;
+
+  /**
+   * Render a custom expand icon for SubMenu components
+   */
   renderExpandIcon?: (params: RenderExpandIconParams) => React.ReactNode;
 }
 
@@ -55,6 +69,8 @@ const StyledMenu = styled.nav<Pick<MenuProps, 'rootStyles'>>`
 `;
 
 export const MenuContext = React.createContext<MenuContextProps | undefined>(undefined);
+
+export const LevelContext = React.createContext<number>(0);
 
 const MenuFR: React.ForwardRefRenderFunction<HTMLMenuElement, MenuProps> = (
   {
@@ -76,14 +92,16 @@ const MenuFR: React.ForwardRefRenderFunction<HTMLMenuElement, MenuProps> = (
 
   return (
     <MenuContext.Provider value={providerValue}>
-      <StyledMenu
-        ref={ref}
-        className={classnames(menuClasses.root, className)}
-        rootStyles={rootStyles}
-        {...rest}
-      >
-        <StyledUl>{children}</StyledUl>
-      </StyledMenu>
+      <LevelContext.Provider value={0}>
+        <StyledMenu
+          ref={ref}
+          className={classnames(menuClasses.root, className)}
+          rootStyles={rootStyles}
+          {...rest}
+        >
+          <StyledUl>{children}</StyledUl>
+        </StyledMenu>
+      </LevelContext.Provider>
     </MenuContext.Provider>
   );
 };